Attempt using sandboxed module... (read on)
I think this will work... next steps are to move all the WebInspector
calls to inside of the sm.require(). It's failing now because the
peekLast is using the outer context when the Array.prototype on the
inner context has been updated.

I'm curious if there could be a wrapper module.. if so, i feel like all
WebInspector calls would have to be WebInspector().

Anyhow. the fix inside this module is moving more over the sm.require()
border. seems ok.